Updated to a minimum working example for a slide out on the model.
The items that dissapear will slide out to the left when they are filtered out.
This works by using a timer in the Rust code before calling filter_model.reset()
essentially when the hide property gets toggled the animation gets triggered in Slint
which lasts for 250s. When the property gets toggled a 500ms timer starts in Rust,
and when that timer is done it calls filter_model.reset(). This works perfectly for
items that are going away.

However so far none of my attempts work on getting something similar for items that
are showing back up. No matter what I have tried they always just appear.