So I'm managed to get a slide in working.
This is really hacky, but it works on a similar principle as the slide
out however it requires an extra property to animate because I have to
set it from the rust code.

Essentially, for the slide in I set the property to be off screen.
Then that gets applied to the X value of the obejcts (but I need to make
care to only apply it to the right properties in the state setting in
Slint). Then reset the filter model, then start a timer, then update
the position, so it animates in.